Fix model placement
